Move tests to the utils package
