+++ packages/mui-material/src/utils/isMuiElement.test.js
@@ -1,0 +1,29 @@
+import * as React from 'react';
+import { expect } from 'chai';
+
+import { isMuiElement } from '.';
+import { Input, ListItemSecondaryAction, SvgIcon } from '..';
+
+describe('utils/index.js', () => {
+  describe('isMuiElement', () => {
+    it('should match static muiName property', () => {
+      const Component = () => null;
+      Component.muiName = 'Component';
+
+      expect(isMuiElement(<Component />, ['Component'])).to.equal(true);
+      expect(isMuiElement(<div />, ['Input'])).to.equal(false);
+      expect(isMuiElement(null, ['SvgIcon'])).to.equal(false);
+      expect(isMuiElement('TextNode', ['SvgIcon'])).to.equal(false);
+    });
+
+    it('should be truthy for matching components', () => {
+      [
+        [Input, 'Input'],
+        [ListItemSecondaryAction, 'ListItemSecondaryAction'],
+        [SvgIcon, 'SvgIcon'],
+      ].forEach(([Component, muiName]) => {
+        expect(isMuiElement(<Component />, [muiName])).to.equal(true);
+      });
+    });
+  });
+});
